Organize Serializer concerns into concerns/
