+++ lib/active_model/serializer.rb
@@ -4,15 +4,15 @@ require 'active_model/serializer/collection_serializer'
 require 'active_model/serializer/array_serializer'
 require 'active_model/serializer/error_serializer'
 require 'active_model/serializer/errors_serializer'
-require 'active_model/serializer/associations'
-require 'active_model/serializer/attributes'
-require 'active_model/serializer/caching'
-require 'active_model/serializer/configuration'
+require 'active_model/serializer/concerns/associations'
+require 'active_model/serializer/concerns/attributes'
+require 'active_model/serializer/concerns/caching'
+require 'active_model/serializer/concerns/configuration'
+require 'active_model/serializer/concerns/links'
+require 'active_model/serializer/concerns/meta'
+require 'active_model/serializer/concerns/type'
 require 'active_model/serializer/fieldset'
 require 'active_model/serializer/lint'
-require 'active_model/serializer/links'
-require 'active_model/serializer/meta'
-require 'active_model/serializer/type'
 
 # ActiveModel::Serializer is an abstract class that is
 # reified when subclassed to decorate a resource.
