+++ lib/active_model/serializer/concerns/attributes.rb
