+++ lib/active_model/serializer/concerns/configuration.rb
