+++ lib/active_model/serializer/concerns/links.rb
