+++ lib/active_model/serializer/concerns/type.rb
